elementwise extension
