+++ ivy/functional/backends/numpy/extensions.py
@@ -59,11 +59,6 @@ def native_sparse_array_to_indices_values_and_shape(x):
     return None, None, None
 
 
-@_handle_0_dim_output
-def sinc(x: np.ndarray, /, *, out: Optional[np.ndarray] = None) -> np.ndarray:
-    return np.sinc(x).astype(x.dtype)
-
-
 def vorbis_window(
     window_length: np.ndarray,
     *,
@@ -85,25 +80,6 @@ def vorbis_window(
 
 
 vorbis_window.support_native_out = False
-
-
-def lcm(
-    x1: np.ndarray,
-    x2: np.ndarray,
-    /,
-    *,
-    out: Optional[np.ndarray] = None,
-) -> np.ndarray:
-    return np.abs(
-        np.lcm(
-            x1,
-            x2,
-            out=out,
-        )
-    )
-
-
-lcm.support_native_out = True
 
 
 def hann_window(
@@ -325,42 +301,3 @@ def flipud(
 
 
 flipud.support_native_out = False
-
-
-def fmod(
-    x1: np.ndarray,
-    x2: np.ndarray,
-    /,
-    *,
-    out: Optional[np.ndarray] = None,
-) -> np.ndarray:
-    return np.fmod(
-        x1,
-        x2,
-        out=None,
-    )
-
-
-fmod.support_native_out = True
-
-
-def fmax(
-    x1: np.ndarray,
-    x2: np.ndarray,
-    /,
-    *,
-    out: Optional[np.ndarray] = None,
-) -> np.ndarray:
-    return np.fmax(
-        x1,
-        x2,
-        out=None,
-        where=True,
-        casting="same_kind",
-        order="K",
-        dtype=None,
-        subok=True,
-    )
-
-
-fmax.support_native_out = True
