+++ ivy/functional/backends/numpy/extensions/elementwise.py
@@ -1,0 +1,75 @@
+from typing import Optional, Union, Tuple, Sequence, Callable, Literal
+from numbers import Number
+import logging
+import ivy
+import numpy as np
+from ivy.functional.ivy.extensions import (
+    _verify_coo_components,
+    _verify_csr_components,
+    _is_coo_not_csr,
+)
+from ivy.functional.backends.numpy.helpers import _handle_0_dim_output
+import math
+
+
+@_handle_0_dim_output
+def sinc(x: np.ndarray, /, *, out: Optional[np.ndarray] = None) -> np.ndarray:
+    return np.sinc(x).astype(x.dtype)
+
+
+def lcm(
+    x1: np.ndarray,
+    x2: np.ndarray,
+    /,
+    *,
+    out: Optional[np.ndarray] = None,
+) -> np.ndarray:
+    return np.abs(
+        np.lcm(
+            x1,
+            x2,
+            out=out,
+        )
+    )
+
+
+lcm.support_native_out = True
+
+
+def fmod(
+    x1: np.ndarray,
+    x2: np.ndarray,
+    /,
+    *,
+    out: Optional[np.ndarray] = None,
+) -> np.ndarray:
+    return np.fmod(
+        x1,
+        x2,
+        out=None,
+    )
+
+
+fmod.support_native_out = True
+
+
+def fmax(
+    x1: np.ndarray,
+    x2: np.ndarray,
+    /,
+    *,
+    out: Optional[np.ndarray] = None,
+) -> np.ndarray:
+    return np.fmax(
+        x1,
+        x2,
+        out=None,
+        where=True,
+        casting="same_kind",
+        order="K",
+        dtype=None,
+        subok=True,
+    )
+
+
+fmax.support_native_out = True
