+++ ivy/functional/backends/tensorflow/extensions.py
@@ -65,16 +65,6 @@ def native_sparse_array_to_indices_values_and_shape(x):
     raise ivy.exceptions.IvyException("not a SparseTensor")
 
 
-def sinc(
-    x: Union[tf.Tensor, tf.Variable],
-    /,
-    *,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    tf.experimental.numpy.experimental_enable_numpy_behavior()
-    return tf.cast(tf.experimental.numpy.sinc(x), x.dtype)
-
-
 def vorbis_window(
     window_length: Union[tf.Tensor, tf.Variable],
     *,
@@ -82,25 +72,6 @@ def vorbis_window(
     out: Optional[Union[tf.Tensor, tf.Variable]] = None,
 ) -> Union[tf.Tensor, tf.Variable]:
     return tf.signal.vorbis_window(window_length, dtype=dtype, name=None)
-
-
-def lcm(
-    x1: Union[tf.Tensor, tf.Variable],
-    x2: Union[tf.Tensor, tf.Variable],
-    /,
-    *,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    if [x1.dtype, x2.dtype] == [tf.int8, tf.int8]:
-        dtype = tf.int8
-        x1 = tf.cast(x1, dtype=tf.int16)
-        x2 = tf.cast(x2, dtype=tf.int16)
-    else:
-        dtype = x1.dtype
-    return tf.math.abs(tf.cast(tf.experimental.numpy.lcm(x1, x2), dtype=dtype))
-
-
-lcm.unsupported_dtypes = ("uint8", "uint16", "uint32", "uint64")
 
 
 def hann_window(
@@ -224,29 +195,3 @@ def flipud(
     out: Optional[Union[tf.Tensor, tf.Variable]] = None,
 ) -> Union[tf.Tensor, tf.Variable]:
     return tf.experimental.numpy.flipud(m)
-
-
-def fmod(
-    x1: Union[tf.Tensor, tf.Variable],
-    x2: Union[tf.Tensor, tf.Variable],
-    /,
-    *,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    return tf.math.floormod(x1, x2, name=None)
-
-
-def fmax(
-    x1: Union[tf.Tensor, tf.Variable],
-    x2: Union[tf.Tensor, tf.Variable],
-    /,
-    *,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    x1 = tf.where(tf.math.is_nan(x1), float("inf"), x1)
-    x2 = tf.where(tf.math.is_nan(x1), float("inf"), x2)
-    ret = tf.math.maximum(x1, x2, name=None)
-    return tf.where(tf.math.is_inf(ret), float("nan"))
-
-
-fmax.supported_dtypes = ("blfoat16", "float16", "float32", "float64")
